LayoutTests/css3/flexbox/inline-flexbox-ignore-firstLine.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
    div::first-line { line-height: 100px; }
    p { line-height: 20px; }
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.inline-flexbox')">
<div id=log></div>
<p>This test inline-flexbox item should ignore flex container's firstLine pseudo element.</p>

<div class="inline-flexbox">
    <p data-expected-height=20>The first item.</p>
    <p data-expected-height=20>The second item.</p>
</div>
</body>
</html>
